requirements.txt
numpy
pandas
matplotlib
Pillow
torch
torchvision
scikit-learn
seaborn
tqdm